feat: support DataFusion 52
Bump datafusion 51 → 52 + datafusion-federation 0.4.12 → 0.5.1.

datafusion-federation pinned to '=0.5.1' explicitly. A loose '^0.5.1'
lets the resolver pull in 0.5.3, which transitively requires DataFusion
53 + arrow 58 and creates a two-DataFusion graph that breaks compile
with confusing 'expected Schema, found Schema' errors.

Three real changes follow from the bump:

1. SQLExecutor::execute() gained a 4th parameter,
   '&[Arc<dyn PhysicalExpr>] filters', in datafusion-federation 0.5.x.
   SqlTable's impl ignores them (pushdown still happens via the
   optimizer rule); ClickHouseTableProvider's impl forwards them to
   the inner reader so future pushdown work can use them. One
   tests/e2e.rs call site also needed the new param.

2. Snapshot fixtures in analyzer::function_pushdown refreshed.
   They had been pre-emptively rewritten to expect 'Int64' literal
   widening for power(int, int), but DataFusion 52 keeps the Float64
   widening that DF51 produced — so the fixtures were wrong both before
   and after this bump. Two-character correction.

3. A latent bug surfaces under DF52 (src/stream.rs):
   When DataFusion requests a 0-column projection (e.g. for COUNT(*),
   where it only needs row counts), project_schema_safe() correctly
   returns Schema::empty() on the DataFusion side, but the streaming
   fast path was passing through ClickHouse batches unchanged.
   ClickHouse can't return zero columns — the unparser still emits
   'SELECT 1 FROM table' as a placeholder — so the underlying stream
   yields batches with one column even though self.schema advertises
   none. DataFusion 51 silently tolerated the mismatch; DataFusion 52
   tightened it to a hard assert_eq! in
   repartition::PerPartitionStream::poll_next_and_coalesce that
   panics the runtime instead of returning an error (arrow-rs#9389).

   Fix: when self.schema is empty, strip the placeholder columns and
   rebuild a 0-column RecordBatch that preserves the row count via
   RecordBatchOptions::with_row_count. RecordBatch defaults to zero
   rows for a column-less batch otherwise, which would silently zero
   out COUNT(*) results.

   This fix restores aggregations, insert_metrics, and sink_write_all
   on DF52.

Also moved one sub-test out of test_clickhouse_udfs_failing because
it no longer fits there:

The 'aggregation over clickhouse function results AND mixed functions'
sub-test (comment: 'NOTE: This fails because of mixed functions.
Needs to be resolved') asserted the query should fail. On DataFusion
52 it now plans correctly — the rewriter isolates the clickhouse
pushdown portion of the projection from the local-only STRING_AGG,
the local aggregation runs cleanly, and the final result is a standard
grouped aggregation:

  +--------+-------+-------------------+-----------+
  | id_mod | total | max_exp           | all_names |
  +--------+-------+-------------------+-----------+
  | 0      | 1     | 7.389056098924109 | Bob       |
  | 1      | 1     | 2.718281828460626 | Alice     |
  +--------+-------+-------------------+-----------+

Promoted it to its own positive-case e2e test
(test_clickhouse_udfs_mixed_functions) so test_clickhouse_udfs_failing
goes back to only holding genuine failure cases. Also updated the
test_complex_agg TODO comment in analyzer::function_pushdown to reflect
that this case now works end-to-end at runtime even though the analyzer
still passes it through as a single un-separated node.

examples/run_all.sh tweak — split build and run, time each separately:
under the old 'cargo run' wrapper, a slow release compile looked
identical to a hung example because all output was redirected to
/dev/null. Now reports as e.g.
'✅ PASS: 02_aggregations (run 6s, build 0s)' so a cold-cache build
is obviously not a hang.

Tests: 'just test' is clean across all six feature-flag combos —
197 + 195 + 172 + 170 + 14 + 16 = 764 tests, zero failures. 'just
checks' is also clean including the full examples/run_all.sh sweep
(12/12 examples pass).

+++ src/analyzer/function_pushdown.rs
@@ -1351,9 +1351,11 @@ mod tests {
         Ok(())
     }
 
-    // TODO: This plan represents a feature that needs to be implemented: how to handle "mixed"
-    // functions in a plan node. Ideally the functions would be separated and the clickhouse
-    // function lowered, but that will take quite a bit of logic.
+    // TODO: This plan still passes the clickhouse function and pure-DataFusion
+    // aggregates through as a single un-separated node. Under DataFusion 52
+    // this produces correct runtime results (see e2e::test_clickhouse_udfs_failing),
+    // but the ideal would still be to separate the clickhouse pushdown from the
+    // local-only aggregates so each side runs where it's most efficient.
     #[cfg(feature = "mocks")]
     #[tokio::test]
     async fn test_complex_agg() -> Result<()> {
@@ -1370,7 +1372,7 @@ mod tests {
         {
             let expected_plan = r#"
             Projection: clickhouse(pow(t.id,Int64(2)),Utf8("Int32")) AS id_mod, count(t.id) AS total, max(clickhouse(exp(t.id),Utf8("Float64"))) AS max_exp
-              Aggregate: groupBy=[[clickhouse(power(CAST(t.id AS Int64), Int64(2)) AS pow(t.id, Int64(2)), Utf8("Int32"))]], aggr=[[count(t.id), max(clickhouse(exp(CAST(t.id AS Float64)), Utf8("Float64")))]]
+              Aggregate: groupBy=[[clickhouse(power(CAST(t.id AS Float64), Float64(2)) AS pow(t.id, Int64(2)), Utf8("Int32"))]], aggr=[[count(t.id), max(clickhouse(exp(CAST(t.id AS Float64)), Utf8("Float64")))]]
                 SubqueryAlias: t
                   TableScan: table2 projection=[id]
             "#
@@ -1407,7 +1409,7 @@ mod tests {
             Union
               Projection: table1.col1 AS id, clickhouse(exp(CAST(table1.col1 AS Float64)), Utf8("Float64")) AS func_id
                 TableScan: table1 projection=[col1], full_filters=[table1.col1 = Int32(1)]
-              Projection: table2.id, clickhouse(power(CAST(table2.id AS Int64), Int64(2)) AS pow(table2.id,Int64(2)), Utf8("Float64")) AS func_id
+              Projection: table2.id, clickhouse(power(CAST(table2.id AS Float64), Float64(2)) AS pow(table2.id,Int64(2)), Utf8("Float64")) AS func_id
                 TableScan: table2 projection=[id], full_filters=[table2.id = Int32(1)]
             "#
             .trim();
